refactor : category model-domain 분리
