+++ blackWhiteDeliveryDriver/src/main/java/com/sparta/blackwhitedeliverydriver/category/service/CategoryService.java
@@ -3,10 +3,10 @@ package com.sparta.blackwhitedeliverydriver.category.service;
 import com.sparta.blackwhitedeliverydriver.category.dto.CategoryIdResponseDto;
 import com.sparta.blackwhitedeliverydriver.category.dto.CategoryRequestDto;
 import com.sparta.blackwhitedeliverydriver.category.dto.CategoryResponseDto;
-import com.sparta.blackwhitedeliverydriver.category.entity.Category;
-import com.sparta.blackwhitedeliverydriver.user.entity.User;
+import com.sparta.blackwhitedeliverydriver.model.category.entity.Category;
+import com.sparta.blackwhitedeliverydriver.model.user.entity.User;
 import com.sparta.blackwhitedeliverydriver.category.exception.CategoryExceptionMessage;
-import com.sparta.blackwhitedeliverydriver.category.repository.CategoryRepository;
+import com.sparta.blackwhitedeliverydriver.model.category.repository.CategoryRepository;
 import lombok.RequiredArgsConstructor;
 import org.springframework.data.domain.Page;
 import org.springframework.data.domain.PageRequest;
